+++ src/index.css
@@ -3,6 +3,6 @@
     padding: 0;
 }
 
-/* body {
+body {
     background-color: #171A21;
-} */
+}
